Restore language-specific cleaning helpers
